% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCflux.R
\name{calcFlux}
\alias{calcFlux}
\title{This function calculates gas fluxes from GC sample data}
\usage{
calcFlux(filelist)
}
\arguments{
\item{filelist}{A file listing the Agilent GC output files to be processed.}
}
\description{
This function converts Agilent GC output.
}
\examples{
calcFlux("./filelist.txt")
}
\seealso{
\code{\link{convert_GC_output}} for the higher-level function which calls this.
}
\keyword{GC}
\keyword{input}
