% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCflux.R
\name{excelToCsv}
\alias{excelToCsv}
\title{This function calculates excelToCsv}
\usage{
excelToCsv(file_path, keep_sheets = NULL, target_dir = NULL, ...)
}
\arguments{
\item{file_path}{A file listing the Agilent GC output files to be processed.}

\item{keep_sheets}{Keep the sheets or not}

\item{target_dir}{Directory to save csv file in}

\item{...}{Further arguments}
}
\description{
This function converts excelToCsv.
}
\seealso{
\code{\link{convert_GC_output}} for the higher-level function which calls this.
}
\keyword{GC}
\keyword{input}
