% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCflux.R
\docType{package}
\name{RCflux-package}
\alias{RCflux}
\alias{RCflux-package}
\title{RCflux calculates fluxes from closed chamber data}
\description{
Gas concentration data may come from a GC or other gas analysers
Data formats for several sources can be read in:
[list here] 
Output is written to CSV and PNG files.
Calculates gas fluxes from closed chamber data.
}
\details{
The only function you're likely to need from \pkg{RCflux} is
\code{\link{calcFlux}}. Refer to the vignettes for details
of how to use it - use \code{vignette()}.
}
\author{
\strong{Maintainer}: Peter Levy \email{plevy@ceh.ac.uk}

}
