% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCflux.R
\name{convert_GC_output}
\alias{convert_GC_output}
\title{Converts Agilent GC output to a standard format}
\usage{
convert_GC_output(filelistIn, dataSource = "aGC")
}
\arguments{
\item{filelistIn}{A file listing the Agilent GC output files to be processed.}

\item{dataSource}{A character string identifying the source of the data: "aGC" from the Agilent GC, or "SRUC_GC" from the SRUC_GC.}
}
\description{
This function converts Agilent GC output.
}
\examples{
convert_GC_output("f:/0Peter/misc/stats/GCflux/filelist_aGCxls.txt")
convert_GC_output("S:/DISE_Instrumentation/GC/Agilent/processing/filelist_aGCxls.txt")
}
\seealso{
\code{\link{calcFlux}} for the higher-level function which calls this.
}
\keyword{GC}
\keyword{input}
